added a time-based access strategy
