+++ cas-server-core/src/test/java/org/jasig/cas/services/DefaultRegisteredServiceAccessStrategyTests.java
@@ -20,7 +20,6 @@
 package org.jasig.cas.services;
 
 import com.google.common.collect.Sets;
-import org.joda.time.DateTime;
 import org.junit.Test;
 
 import java.util.Arrays;
@@ -142,35 +141,6 @@ public class DefaultRegisteredServiceAccessStrategyTests {
         pAttrs.put("cn", "CAS");
         pAttrs.put("givenName", "kaz");
         assertFalse(authz.doPrincipalAttributesAllowServiceAccess(pAttrs));
-    }
-
-    @Test
-    public void checkAuthorizationByRangePass() {
-        final DefaultRegisteredServiceAccessStrategy authz =
-                new DefaultRegisteredServiceAccessStrategy(true, true);
-        authz.setStartingDateTime(DateTime.now().toString());
-        authz.setEndingDateTime(DateTime.now().plusMinutes(10).toString());
-        assertTrue(authz.isServiceAccessAllowed());
-
-    }
-
-    @Test
-    public void checkAuthorizationByRangeFailStartTime() {
-        final DefaultRegisteredServiceAccessStrategy authz =
-                new DefaultRegisteredServiceAccessStrategy(true, true);
-        authz.setStartingDateTime(DateTime.now().plusDays(1).toString());
-        authz.setEndingDateTime(DateTime.now().plusMinutes(10).toString());
-        assertFalse(authz.isServiceAccessAllowed());
-
-    }
-
-    @Test
-    public void checkAuthorizationByRangePassEndTime() {
-        final DefaultRegisteredServiceAccessStrategy authz =
-                new DefaultRegisteredServiceAccessStrategy(true, true);
-        authz.setStartingDateTime(DateTime.now().toString());
-        authz.setEndingDateTime(DateTime.now().plusSeconds(30).toString());
-        assertTrue(authz.isServiceAccessAllowed());
     }
 
     private Map<String, Set<String>> getRequiredAttributes() {
