+++ cas-server-core/src/test/java/org/jasig/cas/services/TimeBasedRegisteredServiceAccessStrategyTests.java
@@ -1,0 +1,63 @@
+/*
+ * Licensed to Apereo under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work
+ * for additional information regarding copyright ownership.
+ * Apereo licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License.  You may obtain a
+ * copy of the License at the following location:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.jasig.cas.services;
+
+import org.joda.time.DateTime;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * The {@link TimeBasedRegisteredServiceAccessStrategyTests} is responsible for
+ * running test cases for {@link TimeBasedRegisteredServiceAccessStrategy}.
+ *
+ * @author Misagh Moayyed
+ */
+public class TimeBasedRegisteredServiceAccessStrategyTests {
+
+    @Test
+    public void checkAuthorizationByRangePass() {
+        final TimeBasedRegisteredServiceAccessStrategy authz =
+                new TimeBasedRegisteredServiceAccessStrategy(true, true);
+        authz.setStartingDateTime(DateTime.now().toString());
+        authz.setEndingDateTime(DateTime.now().plusMinutes(10).toString());
+        assertTrue(authz.isServiceAccessAllowed());
+
+    }
+
+    @Test
+    public void checkAuthorizationByRangeFailStartTime() {
+        final TimeBasedRegisteredServiceAccessStrategy authz =
+                new TimeBasedRegisteredServiceAccessStrategy(true, true);
+        authz.setStartingDateTime(DateTime.now().plusDays(1).toString());
+        authz.setEndingDateTime(DateTime.now().plusMinutes(10).toString());
+        assertFalse(authz.isServiceAccessAllowed());
+
+    }
+
+    @Test
+    public void checkAuthorizationByRangePassEndTime() {
+        final TimeBasedRegisteredServiceAccessStrategy authz =
+                new TimeBasedRegisteredServiceAccessStrategy(true, true);
+        authz.setStartingDateTime(DateTime.now().toString());
+        authz.setEndingDateTime(DateTime.now().plusSeconds(30).toString());
+        assertTrue(authz.isServiceAccessAllowed());
+    }
+}
